+++ utils/src/main/lombok/de/larssh/utils/ResourcePathException.java
@@ -1,0 +1,21 @@
+package de.larssh.utils;
+
+import de.larssh.utils.text.Strings;
+
+/**
+ * Thrown to indicate that a resource path is invalid.
+ */
+public class ResourcePathException extends RuntimeException {
+	private static final long serialVersionUID = -8994130898686688545L;
+
+	/**
+	 * Constructs a new {@link ResourcePathException} with the given message,
+	 * formatting as described at {@link Strings#format(String, Object...)}.
+	 *
+	 * @param message   the detail message
+	 * @param arguments arguments referenced by format specifiers in {@code message}
+	 */
+	public ResourcePathException(final String message, final Object... arguments) {
+		super(Strings.format(message, arguments), null);
+	}
+}